[IND2] Modify PPS_OK LED definition.
The IND2 platform has different bit field definitions for some LED,
Control and Status registers.

We only use the PPS_OK LED in the kernel driver, so only it was updated,
as it is intended to harmonise the IND2 bit field definitions to be
compatible with the IND1 definitions in the near future.

Some minor whitespace formatting.