[IND] Remove `dma_sync` from ISR, add IOCTLs for `dma_sync`
Using `dma_sync*()` calls is slow (especially on zynq).  It takes ~140ms
to sync 64MB of memory, which is too long for an ISR.

Remove the `dma_sync` call from the ISR and create IOCTLs to sync the
entire DMA allocation, or individual banks, from user space.

This should improve performance of the kernel/system as a whole as the
interrupt wont be starving the CPU from other time critical tasks.

Latency from interrupt to `efd_app` dropped from ~140ms to under 1ms !!